Update README.md
The case can open a little, but not all the way because there is something in between the keyboard and the shell holding the computer together.
The RAM is made by Kingston KVR, and it does not display the amount of storage that it is able to hold. 
This laptop is different from other laptops because it is more bulky and durable, while recent laptops are more of tablet style with a connective keyboard. The new laptop are lightweight and becoming smaller.